backend/requirements.txt
# Indian Education Law Chatbot - Backend Requirements
# Compatible with Python 3.13.5

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Vector Database and Search
faiss-cpu>=1.12.0
sentence-transformers>=3.2.0
transformers>=4.46.0

# Machine Learning & NLP
torch>=2.5.0
numpy>=1.24.0,<2.2.0
scikit-learn==1.5.2
nltk==3.9.1

# Data Processing
pandas==2.2.3
jsonschema==4.23.0

# HTTP and API
httpx==0.27.2
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging and Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Development and Testing
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1

# Database (optional for future use)
sqlalchemy==2.0.36
alembic==1.14.0

# Monitoring and Logging
loguru==0.7.2
rich==13.9.4

# Utilities
tqdm==4.66.6
python-dateutil==2.9.0

# AI/LLM Integration
google-generativeai>=0.8.0

# Additional dependencies for Python 3.13.5
typing-extensions>=4.12.0
setuptools>=75.0.0
